Refactor: move message dtos into separate package
