+++ src/main/java/com/chat/chat_api/chatroom/dto/ChatroomSummaryDTO.java
@@ -2,7 +2,7 @@ package com.chat.chat_api.chatroom.dto;
 
 import java.util.List;
 
-import com.chat.chat_api.message.MessageResponseDTO;
+import com.chat.chat_api.message.dto.MessageResponseDTO;
 
 public record ChatroomSummaryDTO(
     Long roomId,
