+++ src/main/java/com/chat/chat_api/message/dto/MessageDTO.java
@@ -1,4 +1,4 @@
-package com.chat.chat_api.message;
+package com.chat.chat_api.message.dto;
 
 import java.time.LocalDateTime;
